requirements.txt
# brew install sox --with-lame --with-flac --with-libvorbis
git+git://github.com/jonathanmarmor/phase.git@master
Pillow